don't update docs when a publish error occurs

+++ travis-publish.sh
@@ -33,52 +33,57 @@ if [[ $TRAVIS_REPO_SLUG == "grails/grails-data-mapping" && $TRAVIS_PULL_REQUEST 
     ./gradlew publish || EXIT_STATUS=$?
   fi
 
-  ./gradlew allDocs || EXIT_STATUS=$?
+  if [[ $EXIT_STATUS -eq 0 ]]; then
+      ./gradlew allDocs || EXIT_STATUS=$?
 
-  git config --global user.name "$GIT_NAME"
-  git config --global user.email "$GIT_EMAIL"
-  git config --global credential.helper "store --file=~/.git-credentials"
-  echo "https://$GH_TOKEN:@github.com" > ~/.git-credentials
+      git config --global user.name "$GIT_NAME"
+      git config --global user.email "$GIT_EMAIL"
+      git config --global credential.helper "store --file=~/.git-credentials"
+      echo "https://$GH_TOKEN:@github.com" > ~/.git-credentials
 
-  git clone https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git -b gh-pages gh-pages --single-branch > /dev/null
-  cd gh-pages
+      git clone https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git -b gh-pages gh-pages --single-branch > /dev/null
+      cd gh-pages
 
-  # If this is the master branch then update the snapshot
-  if [[ $TRAVIS_BRANCH == 'master' ]]; then
-    mkdir -p snapshot
-    cp -r ../build/docs/. ./snapshot/
+      # If this is the master branch then update the snapshot
+      if [[ $TRAVIS_BRANCH == 'master' ]]; then
+        mkdir -p snapshot
+        cp -r ../build/docs/. ./snapshot/
 
-    git add snapshot/*
+        git add snapshot/*
+      fi
+
+      # If there is a tag present then this becomes the latest
+      if [[ $TRAVIS_TAG =~ ^v[[:digit:]] ]]; then
+        version="$TRAVIS_TAG"
+        version=${version:1}
+        milestone=${version:5}
+        if [[ -n $milestone ]]; then
+          mkdir -p latest
+          cp -r ../build/docs/. ./latest/
+          git add latest/*
+        fi
+
+        majorVersion=${version:0:4}
+        majorVersion="${majorVersion}x"
+
+        mkdir -p "$version"
+        cp -r ../build/docs/. "./$version/"
+        git add "$version/*"
+
+        mkdir -p "$majorVersion"
+        cp -r ../build/docs/. "./$majorVersion/"
+        git add "$majorVersion/*"
+
+      fi
+
+      git commit -a -m "Updating docs for Travis build: https://travis-ci.org/$TRAVIS_REPO_SLUG/builds/$TRAVIS_BUILD_ID"
+      git push origin HEAD
+      cd ..
+      rm -rf gh-pages
+  else
+      echo "Error occured during publishing, skipping docs"
   fi
 
-  # If there is a tag present then this becomes the latest
-  if [[ $TRAVIS_TAG =~ ^v[[:digit:]] ]]; then
-    version="$TRAVIS_TAG"
-    version=${version:1}
-    milestone=${version:5}
-    if [[ -n $milestone ]]; then
-      mkdir -p latest
-      cp -r ../build/docs/. ./latest/
-      git add latest/*
-    fi
-
-    majorVersion=${version:0:4}
-    majorVersion="${majorVersion}x"
-
-    mkdir -p "$version"
-    cp -r ../build/docs/. "./$version/"
-    git add "$version/*"
-
-    mkdir -p "$majorVersion"
-    cp -r ../build/docs/. "./$majorVersion/"
-    git add "$majorVersion/*"
-
-  fi
-
-  git commit -a -m "Updating docs for Travis build: https://travis-ci.org/$TRAVIS_REPO_SLUG/builds/$TRAVIS_BUILD_ID"
-  git push origin HEAD
-  cd ..
-  rm -rf gh-pages
 fi
 
 exit $EXIT_STATUS
